visualization: use 2 colors for attribution
Previously, only 1 color was used to display the attributions in the
visualization. However we need a way to distinguish positive and negative
values better, i.e for attribution we need 2 colors instead of 1.

This patch updates the visu for attribution so that by default positive
attributions are displayed in shades of red, and negative attributions
in shades of blue.

Signed-off-by: Frederic Boisnard <frederic.boisnard@irt-saintexupery.com>